+++ src/client/java/niv/heater/HeaterScreen.java
@@ -7,6 +7,7 @@ import net.minecraft.client.gui.screen.ingame.HandledScreen;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
+import niv.heater.screen.HeaterScreenHandler;
 
 import static niv.heater.Heater.MOD_ID;
 
@@ -26,7 +27,6 @@ public class HeaterScreen extends HandledScreen<HeaterScreenHandler> {
 
     @Override
     public void render(DrawContext context, int mouseX, int mouseY, float delta) {
-        renderBackground(context);
         super.render(context, mouseX, mouseY, delta);
         drawMouseoverTooltip(context, mouseX, mouseY);
     }