+++ src/main/java/niv/heater/Heater.java
@@ -12,7 +12,7 @@ import org.slf4j.LoggerFactory;
 
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
-import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
+import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.fabricmc.fabric.api.object.builder.v1.block.entity.FabricBlockEntityTypeBuilder;
 import net.fabricmc.fabric.api.registry.OxidizableBlocksRegistry;
@@ -20,13 +20,21 @@ import net.minecraft.block.Block;
 import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.item.BlockItem;
 import net.minecraft.item.Item;
-import net.minecraft.item.ItemGroups;
-import net.minecraft.item.Items;
+import net.minecraft.item.ItemGroup;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.resource.featuretoggle.FeatureFlags;
 import net.minecraft.screen.ScreenHandlerType;
+import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
+import niv.heater.block.HeatPipeBlock;
+import niv.heater.block.HeaterBlock;
+import niv.heater.block.OxidizableHeatPipeBlock;
+import niv.heater.block.OxidizableHeaterBlock;
+import niv.heater.block.OxidizableThermostatBlock;
+import niv.heater.block.ThermostatBlock;
+import niv.heater.block.entity.HeaterBlockEntity;
+import niv.heater.screen.HeaterScreenHandler;
 
 public class Heater implements ModInitializer {
 
@@ -54,6 +62,16 @@ public class Heater implements ModInitializer {
     public static final Block WAXED_WEATHERED_HEAT_PIPE_BLOCK;
     public static final Block WAXED_OXIDIZED_HEAT_PIPE_BLOCK;
 
+    public static final Block THERMOSTAT_BLOCK;
+    public static final Block EXPOSED_THERMOSTAT_BLOCK;
+    public static final Block WEATHERED_THERMOSTAT_BLOCK;
+    public static final Block OXIDIZED_THERMOSTAT_BLOCK;
+
+    public static final Block WAXED_THERMOSTAT_BLOCK;
+    public static final Block WAXED_EXPOSED_THERMOSTAT_BLOCK;
+    public static final Block WAXED_WEATHERED_THERMOSTAT_BLOCK;
+    public static final Block WAXED_OXIDIZED_THERMOSTAT_BLOCK;
+
     public static final Item HEATER_ITEM;
     public static final Item EXPOSED_HEATER_ITEM;
     public static final Item WEATHERED_HEATER_ITEM;
@@ -74,15 +92,28 @@ public class Heater implements ModInitializer {
     public static final Item WAXED_WEATHERED_HEAT_PIPE_ITEM;
     public static final Item WAXED_OXIDIZED_HEAT_PIPE_ITEM;
 
+    public static final Item THERMOSTAT_ITEM;
+    public static final Item EXPOSED_THERMOSTAT_ITEM;
+    public static final Item WEATHERED_THERMOSTAT_ITEM;
+    public static final Item OXIDIZED_THERMOSTAT_ITEM;
+
+    public static final Item WAXED_THERMOSTAT_ITEM;
+    public static final Item WAXED_EXPOSED_THERMOSTAT_ITEM;
+    public static final Item WAXED_WEATHERED_THERMOSTAT_ITEM;
+    public static final Item WAXED_OXIDIZED_THERMOSTAT_ITEM;
+
     public static final BlockEntityType<HeaterBlockEntity> HEATER_BLOCK_ENTITY;
 
     public static final ScreenHandlerType<HeaterScreenHandler> HEATER_SCREEN_HANDLER;
+
+    public static final ItemGroup HEATER_GROUP;
 
     static {
         MOD_ID = "heater";
 
         final var heater = "heater";
         final var heatPipe = "heat_pipe";
+        final var thermostat = "thermostat";
 
         final var exposed = "exposed_";
         final var weathered = "weathered_";
@@ -114,11 +145,11 @@ public class Heater implements ModInitializer {
         HEAT_PIPE_BLOCK = Registry.register(Registries.BLOCK, id,
                 new OxidizableHeatPipeBlock(UNAFFECTED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
         EXPOSED_HEAT_PIPE_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(exposed),
-                new OxidizableHeatPipeBlock(UNAFFECTED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+                new OxidizableHeatPipeBlock(EXPOSED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
         WEATHERED_HEAT_PIPE_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(weathered),
-                new OxidizableHeatPipeBlock(UNAFFECTED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+                new OxidizableHeatPipeBlock(WEATHERED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
         OXIDIZED_HEAT_PIPE_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(oxidized),
-                new OxidizableHeatPipeBlock(UNAFFECTED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+                new OxidizableHeatPipeBlock(OXIDIZED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
 
         WAXED_HEAT_PIPE_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(waxed),
                 new HeatPipeBlock(UNAFFECTED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
@@ -128,6 +159,25 @@ public class Heater implements ModInitializer {
                 new HeatPipeBlock(WEATHERED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
         WAXED_OXIDIZED_HEAT_PIPE_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(waxed + oxidized),
                 new HeatPipeBlock(OXIDIZED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+
+        id = id.withPath(thermostat);
+        THERMOSTAT_BLOCK = Registry.register(Registries.BLOCK, id,
+                new OxidizableThermostatBlock(UNAFFECTED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+        EXPOSED_THERMOSTAT_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(exposed),
+                new OxidizableThermostatBlock(EXPOSED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+        WEATHERED_THERMOSTAT_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(weathered),
+                new OxidizableThermostatBlock(WEATHERED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+        OXIDIZED_THERMOSTAT_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(oxidized),
+                new OxidizableThermostatBlock(OXIDIZED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+
+        WAXED_THERMOSTAT_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(waxed),
+                new ThermostatBlock(UNAFFECTED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+        WAXED_EXPOSED_THERMOSTAT_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(waxed + exposed),
+                new ThermostatBlock(EXPOSED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+        WAXED_WEATHERED_THERMOSTAT_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(waxed + weathered),
+                new ThermostatBlock(WEATHERED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
+        WAXED_OXIDIZED_THERMOSTAT_BLOCK = Registry.register(Registries.BLOCK, id.withPrefixedPath(waxed + oxidized),
+                new ThermostatBlock(OXIDIZED, FabricBlockSettings.copyOf(COPPER_BLOCK)));
 
         id = id.withPath(heater);
         HEATER_ITEM = Registry.register(Registries.ITEM, id,
@@ -167,6 +217,25 @@ public class Heater implements ModInitializer {
         WAXED_OXIDIZED_HEAT_PIPE_ITEM = Registry.register(Registries.ITEM, id.withPrefixedPath(waxed + oxidized),
                 new BlockItem(WAXED_OXIDIZED_HEAT_PIPE_BLOCK, new FabricItemSettings()));
 
+        id = id.withPath(thermostat);
+        THERMOSTAT_ITEM = Registry.register(Registries.ITEM, id,
+                new BlockItem(THERMOSTAT_BLOCK, new FabricItemSettings()));
+        EXPOSED_THERMOSTAT_ITEM = Registry.register(Registries.ITEM, id.withPrefixedPath(exposed),
+                new BlockItem(EXPOSED_THERMOSTAT_BLOCK, new FabricItemSettings()));
+        WEATHERED_THERMOSTAT_ITEM = Registry.register(Registries.ITEM, id.withPrefixedPath(weathered),
+                new BlockItem(WEATHERED_THERMOSTAT_BLOCK, new FabricItemSettings()));
+        OXIDIZED_THERMOSTAT_ITEM = Registry.register(Registries.ITEM, id.withPrefixedPath(oxidized),
+                new BlockItem(OXIDIZED_THERMOSTAT_BLOCK, new FabricItemSettings()));
+
+        WAXED_THERMOSTAT_ITEM = Registry.register(Registries.ITEM, id.withPrefixedPath(waxed),
+                new BlockItem(WAXED_THERMOSTAT_BLOCK, new FabricItemSettings()));
+        WAXED_EXPOSED_THERMOSTAT_ITEM = Registry.register(Registries.ITEM, id.withPrefixedPath(waxed + exposed),
+                new BlockItem(WAXED_EXPOSED_THERMOSTAT_BLOCK, new FabricItemSettings()));
+        WAXED_WEATHERED_THERMOSTAT_ITEM = Registry.register(Registries.ITEM, id.withPrefixedPath(waxed + weathered),
+                new BlockItem(WAXED_WEATHERED_THERMOSTAT_BLOCK, new FabricItemSettings()));
+        WAXED_OXIDIZED_THERMOSTAT_ITEM = Registry.register(Registries.ITEM, id.withPrefixedPath(waxed + oxidized),
+                new BlockItem(WAXED_OXIDIZED_THERMOSTAT_BLOCK, new FabricItemSettings()));
+
         id = id.withPath(heater);
         HEATER_BLOCK_ENTITY = Registry.register(Registries.BLOCK_ENTITY_TYPE, id,
                 FabricBlockEntityTypeBuilder.create(HeaterBlockEntity::new,
@@ -178,6 +247,42 @@ public class Heater implements ModInitializer {
 
         HEATER_SCREEN_HANDLER = Registry.register(Registries.SCREEN_HANDLER, id,
                 new ScreenHandlerType<>(HeaterScreenHandler::new, FeatureFlags.VANILLA_FEATURES));
+
+        HEATER_GROUP = FabricItemGroup.builder()
+                .icon(HEATER_ITEM::getDefaultStack)
+                .displayName(Text.translatable("itemGroup.heater.tab"))
+                .entries((context, entries) -> {
+                    entries.add(HEATER_ITEM);
+                    entries.add(EXPOSED_HEATER_ITEM);
+                    entries.add(WEATHERED_HEATER_ITEM);
+                    entries.add(OXIDIZED_HEATER_ITEM);
+
+                    entries.add(WAXED_HEATER_ITEM);
+                    entries.add(WAXED_EXPOSED_HEATER_ITEM);
+                    entries.add(WAXED_WEATHERED_HEATER_ITEM);
+                    entries.add(WAXED_OXIDIZED_HEATER_ITEM);
+
+                    entries.add(THERMOSTAT_ITEM);
+                    entries.add(EXPOSED_THERMOSTAT_ITEM);
+                    entries.add(WEATHERED_THERMOSTAT_ITEM);
+                    entries.add(OXIDIZED_THERMOSTAT_ITEM);
+
+                    entries.add(WAXED_THERMOSTAT_ITEM);
+                    entries.add(WAXED_EXPOSED_THERMOSTAT_ITEM);
+                    entries.add(WAXED_WEATHERED_THERMOSTAT_ITEM);
+                    entries.add(WAXED_OXIDIZED_THERMOSTAT_ITEM);
+
+                    entries.add(HEAT_PIPE_ITEM);
+                    entries.add(EXPOSED_HEAT_PIPE_ITEM);
+                    entries.add(WEATHERED_HEAT_PIPE_ITEM);
+                    entries.add(OXIDIZED_HEAT_PIPE_ITEM);
+
+                    entries.add(WAXED_HEAT_PIPE_ITEM);
+                    entries.add(WAXED_EXPOSED_HEAT_PIPE_ITEM);
+                    entries.add(WAXED_WEATHERED_HEAT_PIPE_ITEM);
+                    entries.add(WAXED_OXIDIZED_HEAT_PIPE_ITEM);
+                }).build();
+        Registry.register(Registries.ITEM_GROUP, id.withPath("tab"), HEATER_GROUP);
     }
 
     @Override
@@ -206,15 +311,13 @@ public class Heater implements ModInitializer {
         OxidizableBlocksRegistry.registerWaxableBlockPair(WEATHERED_HEAT_PIPE_BLOCK, WAXED_WEATHERED_HEAT_PIPE_BLOCK);
         OxidizableBlocksRegistry.registerWaxableBlockPair(OXIDIZED_HEAT_PIPE_BLOCK, WAXED_OXIDIZED_HEAT_PIPE_BLOCK);
 
-        ItemGroupEvents.modifyEntriesEvent(ItemGroups.FUNCTIONAL)
-                .register(content -> content.addBefore(Items.FURNACE,
-                        HEATER_ITEM, EXPOSED_HEATER_ITEM,
-                        WEATHERED_HEATER_ITEM, OXIDIZED_HEATER_ITEM,
-                        WAXED_HEATER_ITEM, WAXED_EXPOSED_HEATER_ITEM,
-                        WAXED_WEATHERED_HEATER_ITEM, WAXED_OXIDIZED_HEATER_ITEM,
-                        HEAT_PIPE_ITEM, EXPOSED_HEAT_PIPE_ITEM,
-                        WEATHERED_HEAT_PIPE_ITEM, OXIDIZED_HEAT_PIPE_ITEM,
-                        WAXED_HEAT_PIPE_ITEM, WAXED_EXPOSED_HEAT_PIPE_ITEM,
-                        WAXED_WEATHERED_HEAT_PIPE_ITEM, WAXED_OXIDIZED_HEAT_PIPE_ITEM));
+        OxidizableBlocksRegistry.registerOxidizableBlockPair(THERMOSTAT_BLOCK, EXPOSED_THERMOSTAT_BLOCK);
+        OxidizableBlocksRegistry.registerOxidizableBlockPair(EXPOSED_THERMOSTAT_BLOCK, WEATHERED_THERMOSTAT_BLOCK);
+        OxidizableBlocksRegistry.registerOxidizableBlockPair(WEATHERED_THERMOSTAT_BLOCK, OXIDIZED_THERMOSTAT_BLOCK);
+
+        OxidizableBlocksRegistry.registerWaxableBlockPair(THERMOSTAT_BLOCK, WAXED_THERMOSTAT_BLOCK);
+        OxidizableBlocksRegistry.registerWaxableBlockPair(EXPOSED_THERMOSTAT_BLOCK, WAXED_EXPOSED_THERMOSTAT_BLOCK);
+        OxidizableBlocksRegistry.registerWaxableBlockPair(WEATHERED_THERMOSTAT_BLOCK, WAXED_WEATHERED_THERMOSTAT_BLOCK);
+        OxidizableBlocksRegistry.registerWaxableBlockPair(OXIDIZED_THERMOSTAT_BLOCK, WAXED_OXIDIZED_THERMOSTAT_BLOCK);
     }
 }